# openapi/admin-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chaos Admin API",
    "version": "1.0.0",
    "description": "Passwordless administration API for merchant accounts and headless commerce stores."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
...
  "paths": {
...
    "/admin/v1/stores/{store_id}/payment-provider-accounts": {
      "parameters": [
        {
          "$ref": "#/components/parameters/StoreId"
        }
      ],
      "get": {
        "operationId": "listPaymentProviderAccounts",
        "tags": [
          "Payments"
        ],
        "summary": "List Store Payment Provider accounts without secret references",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment Provider account page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentProviderAccountCollectionEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      },
      "post": {
        "operationId": "createPaymentProviderAccount",
        "tags": [
          "Payments"
        ],
        "summary": "Configure one Payment Provider for a Store",
        "description": "Only owners and administrators may configure providers. Secret-manager references are write-only and are never returned.",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentProviderAccount"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment Provider account configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentProviderAccountEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
    "/admin/v1/stores/{store_id}/payment-provider-accounts/{provider_account_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/StoreId"
        },
        {
          "name": "provider_account_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getPaymentProviderAccount",
        "tags": [
          "Payments"
        ],
        "summary": "Get a Store Payment Provider account without secret references",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payment Provider account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentProviderAccountEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updatePaymentProviderAccount",
        "tags": [
          "Payments"
        ],
        "summary": "Rotate secret references or enable or disable a Payment Provider account",
        "description": "Provider identity and external account mapping are immutable. Secret-manager references are write-only.",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentProviderAccount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment Provider account updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentProviderAccountEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
    "/admin/v1/stores/{store_id}/payment-attempts/{payment_attempt_id}/refunds": {
      "post": {
        "operationId": "createRefund",
        "tags": [
          "Payments"
        ],
        "summary": "Create a Refund against a captured Payment Attempt",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/StoreId"
          },
          {
            "$ref": "#/components/parameters/PaymentAttemptId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRefund"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Refund created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque revocable session token"
      }
    },
    "parameters": {
      "StoreId": {
        "name": "store_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "PaymentAttemptId": {
        "name": "payment_attempt_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque cursor returned by the previous page.",
        "schema": {
          "type": "string"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed JSON, query, content type, or payload size",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or credential invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated caller is not allowed to access the resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist in the authorized scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency or uniqueness conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "Domain or request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "CreateRefund": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "amount_minor"
        ],
        "properties": {
          "amount_minor": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      },
      "CreatePaymentProviderAccount": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "provider",
          "display_name",
          "external_account_reference",
          "credential_secret_reference",
          "webhook_secret_reference",
          "enabled"
        ],
        "properties": {
          "provider": {
            "type": "string",
            "pattern": "^[a-z0-9_]{1,64}$"
          },
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "external_account_reference": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "credential_secret_reference": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "writeOnly": true,
            "description": "Opaque reference resolved by the deployment secret manager; never raw credentials."
          },
          "webhook_secret_reference": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "writeOnly": true,
            "description": "Opaque reference resolved by the deployment secret manager; never a raw signing secret."
          },
          "enabled": {
            "type": "boolean",
            "description": "Requests a live Provider readiness check; effective enablement requires a ready result."
          }
        }
      },
      "UpdatePaymentProviderAccount": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "display_name",
          "credential_secret_reference",
          "webhook_secret_reference",
          "enabled"
        ],
        "properties": {
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "credential_secret_reference": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "writeOnly": true
          },
          "webhook_secret_reference": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "writeOnly": true
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "PaymentProviderAccount": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "provider",
          "display_name",
          "external_account_reference",
          "enabled",
          "credentials_configured",
          "readiness_status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "pattern": "^[a-z0-9_]{1,64}$"
          },
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "external_account_reference": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "enabled": {
            "type": "boolean"
          },
          "credentials_configured": {
            "type": "boolean"
          },
          "readiness_status": {
            "type": "string",
            "enum": [
              "unchecked",
              "ready",
              "action_required"
            ]
          },
          "readiness_checked_at": {
            "type": "string",
            "format": "date-time"
          },
          "readiness_valid_until": {
            "type": "string",
            "format": "date-time",
            "description": "End of the current readiness validity window. Enabled accounts are rechecked before this time."
          },
          "readiness_blocker_codes": {
            "type": "array",
            "maxItems": 16,
            "items": {
              "type": "string",
              "pattern": "^[a-z0-9_]{1,64}$"
            }
          },
          "credential_rotation_expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "End of the rollback window for the previous outbound credential reference, when a rotation is active."
          },
          "webhook_rotation_expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "End of the verification overlap window for the previous webhook secret reference, when a rotation is active."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaymentProviderAccountEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaymentProviderAccount"
          }
        }
      },
      "PaymentProviderAccountCollectionEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentProviderAccount"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
      "Refund": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "payment_attempt_id",
          "amount_minor",
          "currency",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "payment_attempt_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount_minor": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "succeeded",
              "failed"
            ]
          },
          "provider_reference": {
            "type": "string"
          },
          "failure_code": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RefundEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Refund"
          }
        }
      },
      "PageMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "next_cursor": {
            "type": "string"
          }
        }
      },
      "ResponseMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "page"
        ],
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    }
  }
}